Extract photon_feedback_derive_tests.rs from turn.rs (parent #680)
Move the four #[cfg(test)] photon-feedback derive unit-test mods
appended to the end of turn.rs into a new
src/agent/loop_run/photon_feedback_derive_tests.rs sibling test file:

- derive_photon_feedback_outcome_tests (Issue #591 Phase 3, cases A-F)
- is_rerun_trigger_tests
- rerun_hint_eligibility_tests
- prepare_adopted_ids_for_evaluate_tests

All four consume `super::super::photon_feedback_derive::*` (sibling
SSOT). #[cfg(test)] only; production binary excludes this file. No
facade re-export (DR3-001). Test paths are preserved unchanged because
the new sibling sits at the same module depth as turn.rs.

turn.rs: 26,438 -> 25,687 (-751 LOC)
Cumulative from 39,489: -13,802 LOC (-35.0%)

Verification:
- cargo fmt --check
- cargo clippy --lib --all-targets -- -D warnings
- cargo test --lib (3,114 passed)
